# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
...
  '/api/favorites/{id}':
    delete:
      summary: 'Eliminar un personaje de la lista de favoritos.'
      operationId: eliminarUnPersonajeDeLaListaDeFavoritos
      description: ''
      parameters:
        -
          in: query
          name: id
          description: 'Id del personaje'
          example: 10
          required: true
          schema:
            type: integer
            description: 'Id del personaje'
            example: 10
      responses:
        202:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: false
                  message: 'Character removed from favorites'
                properties:
                  error:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Character removed from favorites'
        404:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: true
                  message: 'Not found'
                properties:
                  error:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Not found'
      tags:
        - Autenticado
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the favorite.'
        example: nulla
        required: true
        schema:
          type: string
...
  '/api/characters/{id}':
    get:
      summary: 'Mostrar un personaje.'
      operationId: mostrarUnPersonaje
      description: ''
      parameters:
        -
          in: query
          name: id
          description: 'Id del personaje'
          example: 4
          required: false
          schema:
            type: integer
            description: 'Id del personaje'
            example: 4
      responses:
        200:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "{\n   \"id\": integer,\n   \"name\": string,\n   \"status\": string,\n   \"species\": string,\n   \"type\": string,\n   \"gender\": string,\n   \"origin\": {\n       \"name\": string,\n       \"url\": string\n   },\n   \"location\": {\n       \"name\": string,\n       \"url\": string\n   },\n   \"image\": string,\n   \"episode\": [\n       string\n   ],\n   \"url\": string,\n   \"created\": string\n}"
        404:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: true
                  message: 'Character not found'
                properties:
                  error:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Character not found'
      tags:
        - Personajes
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the character.'
        example: ipsum
        required: true
        schema:
          type: string
...
components:
  securitySchemes:
    default:
      type: apiKey
      name: Authorization
      in: header
      description: 'You can retrieve your token by visiting your dashboard and clicking <b>Generate API token</b>.'
security:
  -
    default: []